+++ docs/source/changelog.rst
@@ -7,6 +7,9 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 
 [Unreleased]
 ------------
+Fixed
+^^^^^^^
+* Outdated default GameME link for eGO
 
 [2.2.0] - 2022-11-04
 ---------------------